+++ sync/glue/notes_model_associator.cpp
@@ -938,7 +938,7 @@ syncer::SyncError NotesModelAssociator::CheckModelSyncState(
     } else {
       UMA_HISTOGRAM_ENUMERATION("Sync.LocalModelOutOfSync",
                                 ModelTypeToHistogramInt(syncer::NOTES),
-                                syncer::MODEL_TYPE_COUNT);
+                                static_cast<int>(syncer::MODEL_TYPE_COUNT));
 
       // Clear version on notes model so that we only report error once.
       notes_model_->SetNodeSyncTransactionVersion(
